% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dependencies.R
\name{parse_package_dependencies}
\alias{parse_package_dependencies}
\title{parse package dependencies}
\usage{
parse_package_dependencies(
  packages,
  ap,
  dep_types = c("Depends", "Imports", "Suggests", "LinkingTo"),
  filter_missing = TRUE,
  .df = TRUE
)
}
\arguments{
\item{packages}{vector of package names}

\item{ap}{dataframe-like structure containing dep vectors returned by available.packages()}

\item{dep_types}{vector of dependency types such as Depends, Imports, etc}

\item{filter_missing}{filter packages that have no dependencies}

\item{.df}{whether to return a data frame, default: TRUE}
}
\description{
parse package dependencies
}
\details{
available.packages returns the dependency of a package as a single string that is comma separated
and unparsed containing things like newlines.
}
